+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - exit the program while `delete` command with returned code other than `NOT_ALLOWED` or `CONFIGDB_OBJECT_DEPENDENCY`
 - use java.net.URI instead of URLEncoder.encode, since the latter will translate space into "+" instead of "%20"
 - disable the message-vpn before update attribute `restTlsServerCertValidateNameEnabled`
+- OpaquePassword is only capable when the sempVersion of the config file is same as the broker's.
 
 ## [1.1.1] - 2021-04-06
 ### Added